Move expect types to @jest/types
