+++ packages/jest-types/src/Expect.ts
@@ -1,0 +1,14 @@
+export type SyncExpectationResult = {
+  pass: boolean;
+  message: () => string;
+};
+
+export type AsyncExpectationResult = Promise<SyncExpectationResult>;
+
+export type ExpectationResult = SyncExpectationResult | AsyncExpectationResult;
+
+export type RawMatcherFn = (
+  expected: any,
+  actual: any,
+  options?: any,
+) => ExpectationResult;
