+++ packages/jest-types/src/index.ts
@@ -16,6 +16,7 @@ import * as SourceMaps from './SourceMaps';
 import * as TestResult from './TestResult';
 import * as Global from './Global';
 import * as Environment from './Environment';
+import * as Expect from './Expect';
 
 export {
   Config,
@@ -29,4 +30,5 @@ export {
   TestResult,
   Global,
   Environment,
+  Expect,
 };
